WIP: Add tool to generate grammar from the SyntaxNode APIs
For documentation purposes we need to define a grammar. One option is
to hand write this from the parser. The other option to tool a baseline
from the SyntaxNode APIs.

Ideally, our grammar would be a bit more detailed though, in order to
get precedences and operator priorities across. However, that's hard
to tool because that would require inferring that information from the
structure of our recursive decent parser.